move AppLayout to layouts directory; update tailwind config
