+++ client/src/layouts/AppLayout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import NavBar from "./NavBar";
+import NavBar from "../NavBar";
 
 const AppLayout: React.FC = () => {
   return (
